Pont_Mbale/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==3.2
django-cors-headers==4.4.0
djangorestframework==3.12.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.1.3
fastapi==0.103.0
h11==0.14.0
idna==3.4
importlib_resources==6.4.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
pillow==10.3.0
pkgutil_resolve_name==1.3.10
psycopg2==2.9.9
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.5.0
starlette==0.27.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.23.2
zipp==3.19.2
